Move generated_mxid(), remove server_name()
- Move generate_mxid from Matrix to APIHelper class since server_name is not
  available anymore and restrieve_homeserver_name is located there as well.
- Fix all generate_mxid calls in cli methods.
- Fix and move imports in api.py and cli/__init__.py.
- Make uri arg optional in retrieve_homeserver_name (Fall back to configured
  base_url when nothing is passed).
