+++ synadm/cli/media.py
@@ -85,7 +85,7 @@ def media_list_cmd(ctx, helper, room_id, user_id, from_, limit, sort, reverse,
         helper.output(media_list)
     elif user_id:
         from synadm.cli import user
-        mxid = helper.matrix_api.generate_mxid(user_id)
+        mxid = helper.generate_mxid(user_id)
         ctx.invoke(
             user.get_function("user_media_cmd"),
             user_id=mxid, from_=from_, limit=limit, sort=sort,
@@ -133,7 +133,7 @@ def media_quarantine_cmd(helper, server_name, media_id, user_id, room_id):
     elif room_id:
         media_quarantined = helper.api.room_media_quarantine(room_id)
     elif user_id:
-        mxid = helper.matrix_api.generate_mxid(user_id)
+        mxid = helper.generate_mxid(user_id)
         media_quarantined = helper.api.user_media_quarantine(mxid)
 
     if media_quarantined is None:
