+++ synadm/cli/room.py
@@ -41,7 +41,7 @@ def room():
 def join(helper, room_id_or_alias, user_id):
     """ Join a room
     """
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     out = helper.api.room_join(room_id_or_alias, mxid)
     helper.output(out)
 
@@ -259,7 +259,7 @@ def delete(ctx, helper, room_id, new_room_user_id, room_name, message, block,
                      type=bool, default=False, show_default=False)
     )
     if sure:
-        mxid = helper.matrix_api.generate_mxid(new_room_user_id)
+        mxid = helper.generate_mxid(new_room_user_id)
         room_del = helper.api.room_delete(
             room_id, mxid, room_name,
             message, block, no_purge)
@@ -312,6 +312,6 @@ def make_admin(helper, room_id, user_id):
     the user.
     """
 
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     out = helper.api.room_make_admin(room_id, mxid)
     helper.output(out)
