+++ synadm/cli/user.py
@@ -69,7 +69,7 @@ def user():
 def list_user_cmd(helper, from_, limit, guests, deactivated, name, user_id):
     """ List and search for users
     """
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     users = helper.api.user_list(from_, limit, guests, deactivated, name,
                                  mxid)
     if users is None:
@@ -110,7 +110,7 @@ def deactivate(ctx, helper, user_id, gdpr_erase):
     - Password reset
     - Deletion of third-party-IDs (to prevent the user requesting a password)
     """)
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     ctx.invoke(user_details_cmd, user_id=mxid)
     ctx.invoke(membership, user_id=mxid)
     m_erase_or_deact = "gdpr-erase" if gdpr_erase else "deactivate"
@@ -178,7 +178,7 @@ def prune_devices_cmd(helper, user_id, list_only, min_days, min_surviving,
     Note that this will affect the encryption and decryption of messages sent by
     other users to this user or to rooms where the user is present.
     """
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     devices_data = helper.api.user_devices(mxid)
     if "devices" not in devices_data:
         # Most probably the requested user is not existing, show error and quit.
@@ -230,7 +230,7 @@ def password_cmd(helper, user_id, password, no_logout):
 
     To prevent the user from being logged out of all sessions use option -n.
     """
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     changed = helper.api.user_password(mxid, password, no_logout)
     if changed is None:
         click.echo("Password could not be reset.")
@@ -255,7 +255,7 @@ def membership(helper, user_id, aliases):
 
     Provide matrix user ID (@user:server) as argument.
     """
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     joined_rooms = helper.api.user_membership(mxid, aliases,
                                               helper.matrix_api)
     if joined_rooms is None:
@@ -303,7 +303,7 @@ def user_search_cmd(ctx, search_term, from_, limit):
 def user_details_cmd(helper, user_id):
     """ View details of a user account.
     """
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     user_data = helper.api.user_details(mxid)
     if user_data is None:
         click.echo("User details could not be fetched.")
@@ -373,7 +373,7 @@ def modify(ctx, helper, user_id, password, password_prompt, display_name,
             "Deactivating a user and setting a password doesn't make sense.")
         raise SystemExit(1)
 
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     click.echo("Current user account settings:")
     ctx.invoke(user_details_cmd, user_id=mxid)
     click.echo("User account settings to be modified:")
@@ -432,7 +432,7 @@ def modify(ctx, helper, user_id, password, password_prompt, display_name,
 def whois(helper, user_id):
     """ Return information about the active sessions for a specific user
     """
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     user_data = helper.api.user_whois(mxid)
     helper.output(user_data)
 
@@ -479,7 +479,7 @@ def user_media_cmd(helper, user_id, from_, limit, sort, reverse, datetime):
     safe_from_quarantine), this can cause a large load on the database,
     especially for large environments
     """
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     media = helper.api.user_media(mxid, from_, limit, sort, reverse,
                                   datetime)
     if media is None:
@@ -541,7 +541,7 @@ def user_login_cmd(helper, user_id, expire_days, expire, expire_ts,
     admin user calls /logout/all from any of their devices, but the token will
     not expire if the target user does the same.
     """
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     if expire_never:
         user_login = helper.api.user_login(mxid, None, None, None)
     elif not expire_days and not expire and not expire_ts:
@@ -583,7 +583,7 @@ def user_shadow_ban(helper, user_id, unban):
 
     Generally, it is more appropriate to ban or kick abusive users.
     """
-    mxid = helper.matrix_api.generate_mxid(user_id)
+    mxid = helper.generate_mxid(user_id)
     user_ban = helper.api.user_shadow_ban(mxid, unban)
     if user_ban is None:
         click.echo("Failed to shadow-ban: {}".format(user_id))
